Applied FTB update and release new version (2.2.0)
FTB V1.4.11

+++ src/main/java/net/ftb/tools/ModManager.java
@@ -16,6 +16,23 @@
  */
 package net.ftb.tools;
 
+import static com.google.common.net.HttpHeaders.CACHE_CONTROL;
+import static com.google.common.net.HttpHeaders.CONTENT_MD5;
+import static net.ftb.download.Locations.MODPACKS;
+import static net.ftb.download.Locations.PRIVATEPACKS;
+
+import net.ftb.data.ModPack;
+import net.ftb.data.Settings;
+import net.ftb.gui.LaunchFrame;
+import net.ftb.gui.dialogs.ModpackUpdateDialog;
+import net.ftb.log.Logger;
+import net.ftb.util.DownloadUtils;
+import net.ftb.util.ErrorUtils;
+import net.ftb.util.FTBFileUtils;
+import net.ftb.util.ModPackUtil;
+import net.ftb.util.OSUtils;
+import net.ftb.util.TrackerUtils;
+
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.io.BufferedInputStream;
@@ -29,28 +46,12 @@ import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.concurrent.ExecutionException;
 
-import javax.swing.JDialog;
-import javax.swing.JFrame;
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-import javax.swing.JProgressBar;
-import javax.swing.SwingConstants;
-import javax.swing.SwingUtilities;
-import javax.swing.SwingWorker;
+import javax.swing.*;
 import javax.swing.border.EmptyBorder;
 
-import net.ftb.data.ModPack;
-import net.ftb.data.Settings;
-import net.ftb.gui.LaunchFrame;
-import net.ftb.gui.dialogs.ModpackUpdateDialog;
-import net.ftb.log.Logger;
 import net.ftb.util.AppUtils;
-import net.ftb.util.*;
-import net.ftb.util.FTBFileUtils;
-
-import static net.ftb.download.Locations.MODPACKS;
-import static net.ftb.download.Locations.PRIVATEPACKS;
 
 @SuppressWarnings("serial")
 public class ModManager extends JDialog {
@@ -75,6 +76,8 @@ public class ModManager extends JDialog {
                     File modPackZip = new File(installPath, "ModPacks" + sep + pack.getDir() + sep + pack.getUrl());
                     if (modPackZip.exists()) {
                         FTBFileUtils.delete(modPackZip);
+                        //Also clear out the "default mods" cache entry, if any, to force it to update when next requested
+                        ModPackUtil.clearDefaultModFiles(pack);
                     }
                     File animationGif = new File(OSUtils.getCacheStorageLocation(), "ModPacks" + sep + pack.getDir() + sep + pack.getAnimation());
                     if (animationGif.exists()) {
@@ -91,7 +94,8 @@ public class ModManager extends JDialog {
             return true;
         }
 
-        public String downloadUrl (String filename, String urlString) {
+        public String downloadUrl (String filename, String urlString) throws Exception{
+            boolean failed = false;
             BufferedInputStream in = null;
             FileOutputStream fout;
             HttpURLConnection connection = null;
@@ -108,6 +112,7 @@ public class ModManager extends JDialog {
 
             do {
                 try {
+                    failed = false;
                     startAmount = amount;
 
                     if (amount > 0) {
@@ -123,7 +128,7 @@ public class ModManager extends JDialog {
                     });
 
                     connection = (HttpURLConnection) url_.openConnection();
-                    connection.setRequestProperty("Cache-Control", "no-transform");
+                    connection.setRequestProperty(CACHE_CONTROL, "no-transform");
                     connection.setAllowUserInteraction(true);
                     connection.setConnectTimeout(14000);
                     connection.setReadTimeout(20000);
@@ -131,8 +136,7 @@ public class ModManager extends JDialog {
                         connection.setRequestProperty("Range", "bytes=" + amount + "-");
                     }
                     connection.connect();
-                    ModPack pack = ModPack.getSelectedPack();
-                    md5 = AppUtils.downloadString(new URL("http://feedthenuketerrorist.fr.nf/MD5/" + pack.getDir() + "/" + curVersion + ".md5"));
+                    md5 = AppUtils.downloadString(new URL("http://feedthenuketerrorist.fr.nf/MD5/" + ModPack.getSelectedPack().getDir() + "/" + curVersion + ".md5"));
                     in = new BufferedInputStream(connection.getInputStream());
                     if (modPackSize == 0) {
                         modPackSize = connection.getContentLength();
@@ -147,8 +151,9 @@ public class ModManager extends JDialog {
                     while ((count = in.read(data, 0, 1024)) != -1) {
                         fout.write(data, 0, count);
 
-                        if (count > 0)
+                        if (count > 0) {
                             retryCount = 5;
+                        }
 
                         downloadedPerc += (count * 1.0 / modPackSize) * 100;
                         amount += count;
@@ -167,8 +172,10 @@ public class ModManager extends JDialog {
                     }
                 } catch (MalformedURLException e) {
                     Logger.logError("Error while downloading modpack", e);
+                    failed = true;
                 } catch (IOException e) {
                     Logger.logError("Error while downloading modpack", e);
+                    failed = true;
                 }
 
                 try {
@@ -190,6 +197,10 @@ public class ModManager extends JDialog {
                 }
             } catch (IOException e) {
                 Logger.logWarn("Error while downloading modpack", e);
+            }
+
+            if (failed) {
+                throw new Exception("Modpack download failed");
             }
             return md5;
         }
@@ -224,17 +235,18 @@ public class ModManager extends JDialog {
                     File animationFile = new File(baseDynamic.getPath() + sep + animation);
 
                     if (!animation.equalsIgnoreCase("empty") && !animationFile.exists()) {
-                            downloadUrl(baseDynamic.getPath() + sep + animation, DownloadUtils.getCreeperhostLink(baseLink + animation));
+                        downloadUrl(baseDynamic.getPath() + sep + animation, DownloadUtils.getCreeperhostLink(baseLink + animation));
                     }
                 }
             } catch (Exception e) {
-                Logger.logError("Error validating pack archive", e);
+                ErrorUtils.tossError("Error while downloading modpack", e);
+                return false;
             }
 
             try {
                 if (!dir.equals("mojang_vanilla")
                         && ((md5 == null || md5.isEmpty()) ? DownloadUtils.backupIsValid(new File(baseDynamic, modPackName), baseLink + modPackName) : DownloadUtils.isValid(new File(baseDynamic,
-                                modPackName), md5))) {
+                        modPackName), md5))) {
                     Logger.logDebug("Extracting pack.");
                     Logger.logDebug("Purging mods, coremods, instMods");
                     clearModsFolder(pack);
@@ -254,11 +266,15 @@ public class ModManager extends JDialog {
                     }
 
                     Logger.logDebug("Extracting pack.");
-                    FTBFileUtils.extractZipTo(baseDynamic.getPath() + sep + modPackName, baseDynamic.getPath());
+                    if (!FTBFileUtils.extractZipTo(baseDynamic.getPath() + sep + modPackName, baseDynamic.getPath())) {
+                        ErrorUtils.tossError("Error unzipping modpack!!!");
+                        return false;
+                    }
                     if (pack.getBundledMap() && saveExists) {
                         try {
-                            if (new File(installPath, dir + "/minecraft/saves").exists() && new File(installPath, dir + "/minecraft/saves.ftbtmp").exists())
+                            if (new File(installPath, dir + "/minecraft/saves").exists() && new File(installPath, dir + "/minecraft/saves.ftbtmp").exists()) {
                                 FTBFileUtils.delete(new File(installPath, dir + "/minecraft/saves"));
+                            }
                             if (new File(installPath, dir + "/minecraft/saves.ftbtmp").exists()) {
                                 FTBFileUtils.copyFolder(new File(installPath, dir + "/minecraft/saves.ftbtmp"), new File(installPath, dir + "/minecraft/saves"), true);
                                 FTBFileUtils.delete(new File(installPath, dir + "/minecraft/saves.ftbtmp"));
@@ -297,7 +313,7 @@ public class ModManager extends JDialog {
     /**
      * Create the frame.
      */
-    public ModManager(JFrame owner, Boolean model) {
+    public ModManager (JFrame owner, Boolean model) {
         super(owner, model);
         setResizable(false);
         setTitle("Downloading...");
@@ -327,6 +343,13 @@ public class ModManager extends JDialog {
                 worker = new ModManagerWorker() {
                     @Override
                     protected void done () {
+                        try {
+                            get();
+                        } catch (InterruptedException e) {
+                            Logger.logDebug("Swingworker Exception", e);
+                        } catch (ExecutionException e) {
+                            Logger.logDebug("Swingworker Exception", e.getCause());
+                        }
                         setVisible(false);
                         super.done();
                     }
@@ -430,8 +453,9 @@ public class ModManager extends JDialog {
                 if (file.toLowerCase().endsWith(".zip") || file.toLowerCase().endsWith(".jar") || file.toLowerCase().endsWith(".disabled") || file.toLowerCase().endsWith(".litemod")) {
                     try {
                         boolean b = FTBFileUtils.delete(new File(folder, file));
-                        if (!b)
+                        if (!b) {
                             Logger.logInfo("Error deleting " + file);
+                        }
                     } catch (IOException e) {
                         Logger.logWarn(e.getMessage(), e);
